web/requirements.txt
Jinja2==3.1.5
fastapi-cli==0.0.7
fastapi==0.115.8
httptools==0.6.4
httpx==0.28.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-multipart==0.0.20
uvicorn==0.34.0
